Allocate client buffer & damage only for surfaces that belong to a renderable layer
This is was the behaviour before XDamage implementation. So this
change reverts back to the old behaviour.

A renderable layer is one which is part of getCurrentRenderOrder().
Surfaces which don't belong to such a layer are not rendered
in the final compositing, hence we don't need to allocate/create
client buffer & damage for such surfaces. Instead this is delayed
till the surface is added to a renderable layer.