Feature Patch : moving render specific base implementations to library
This patch opens the possibility to extend the currently provided base
libs for GLES and GLX on vendor platform specific implementations. Therefore
a vendor is able to use the functionality of already existing renderer
approaches for surfaces / layers and extend these functionality to do
platform specific improvements e.g. scaling of a hardware video layer.