Introducion of process function to dispatch Communication Messages
Sometime it is needed that you don't want to run the communicator
in a own thread. Therefore it is needed that you can trigger read
write to the communication message queue inside of the mainloop.

Thatis the reason why we introduce the new process method in the
ICommunicator interface and provides a reference implementation
inside of the DBUSCommunicator.