WindowSystem: Don't composite for every redraw request
Previously composition was done for every redraw request. But this
is incorrect as a layer/surface not part of the current render order
may have sent that request.

Correcting this exposes two hidden bugs. When the render order is
modified, there is no way to compare it against the old render order.
Hence we need to force a complete composition. And updating uniforms
of a shader also needs to force a composition.

Change-Id: Ie255c85fda24f6f30504617f19b62a23d77bbea5