Add input event support for X11WindowSystem
This patch adds support for pointer & keyboard events to X11 window system.
It can serve as example to introduce input events support for other window system.

Note that MultiTouch is _not_ yet supported for the following reasons:
 1- I don't have any plateform runing a recent X11, which is needed for validation
 2- We would need to upgrade the whole X11WindowSystem to use XInput2, & I don't have the time for the moment
I've added comments in X11WindowSystem.cpp to indicate what has to be done. If someone wants to takeover this task .. ;)