Improve plugin search over specific filesystem type
"man readdir" explains that only few filesystem type support a good filling of  "struct dirent.d_type". (nfs does not for instance)
To be POSIX compliant, we should use stat() instead to look if a file is a regular file.

This patch update getSharedLibrariesFromDirectory() to use  stat() when necessary & follow symlinks until reaching the real target.