ilmClient: added internal thread, updated IpcModule API, updated tests
This patch adds an internal thread thread which implements a receive loop.
The internal thread is required to process any incoming notification event
independent of the current client state.

Inside the client API a list of callbacks is stored to support different
callbacks for different layer and surface notification.

The behavior of the API is unchanged, but additionally asynchronous callbacks
can be supported.

add notification tests:
tests all possible individual notification targets for layers
and surfaces and some combinations aswell.

remove destroyMessage:
was only used for DBusIpcModule, but lifecycle of received messages
is now handled internally. The external call is not required any more.

added thread for notification delivery:
the receive thread used to execute notification callbacks directly,
but this allowed a client to block the receive loop.
Now the receive thread adds all received notifications to a queue.
This queue is processed by the new notification thread, which
triggers the callbacks for incoming notifications.
if a client blocks the notification thread, it will only affect incoming
notifications, but the receive thread will continue working.
If the notification queue is full, the receive thread discards incoming
notifications.