GenericCommunicator: added notification support
Updated to use new IpcModule API.

Manages internal list of registered notifications. This list is used
to remove notifications on client disconnect.

After processing an incoming command, the list of client notifications
is processed: each notification is created using the IpcModule and
sent to the appropriate list of clients.

removed destroyMessage:
was only used for DBusIpcModule, but lifecycle of received messages
is now handled internally. The external call is not required any more.