DbusIpcModule: updated to new IpcModule API, added notification support
Send and Receive have been reimplemented to support asynchronous sending
and receiving of events. This enables the use of DBus signals as transport
for client notifications.

The implementation now features a DBus main loop based on DBusWatches,
select(), DBus filters and dbus_connection_dispatch().

Note: Layer- & SurfaceNotification messages are detected and the DBus
signal handlers are setup as required inside the DbusIpcModule.

removed destroyMessage:
was only used for DBusIpcModule, but lifecycle of received messages
is now handled internally. The external call is not required any more.