Renderers: new GLES shader selection mechanism
To accomodate upcoming GLESGraphicSystem optimizations that introduce
many shaders, a new key-based method is introduced for shader
selection.  A "shader key" is an integer containing a bitfield
describing all aspects of a shader, such as whether textures require
uniform transparency multiplication, colorkeying, if it blends with
the framebuffer, and how many textures are referenced in the shader.
The shader key system is overkill for the current set of shaders, but
it is the machinery to allow a large variety of new optimized shaders
to be added without requring any new logic.