GLESGraphicSystem: multitexture and skip clear optimizations
Introduce "multitexture" and "skip clear" optimization support for
GLESGraphicSystem.  The two optimizations are separately configurable,
and may be enabled or disabled independently in the ilm API.

1. Multitexture calculates the screenspace overlapping regions for all
visible surfaces, and then renders each region exactly once, sampling
from the overlapping surfaces in a single shader. Surface rotation
is currently supported, but Layer rotation is not, so this option
is automatically disabled if any layers have non-zero orientations.

2. SkipClear removes the per-frame glClear(), also disables blending for
the first layer of surfaces that are drawn.  Rather than limiting
the optimization to situations where the first surface drawn is full
screen, this optimization will draw additional quads of clear color
to fill the empty screenspace so that glClear() can still be avoided
when benificial.

Both optimizations are geared toward reducing memory bandwidth caused
by reading and writing to the framebuffer several times.  When both
are turned on, and there exists shaders capable of rendering all
screen regions in a single pass, the framebuffer is only written to
once, and never read from.

New shaders:
   Zero surface shader (for partial clears)
   Single surface shaders (opt. for blendless and full opacity)
   Dual surface (also opt. for blendless and full opacity)

All shaders now have "blend pragmas" which are required for NVIDIA
Tegra's binary shader mode.  The pragma's will be safely ignored
on other vendor's hardware, per GLES specification.