LayerManagerPlugins: introduced WITH_STATIC_PLUGINS to statically link plugins
the build system was enhanced to create a list of
statically linked plungins. All plugins in this list are statically linked
against LayerManagerService.

To enable this option, the build flag WITH_STATIC_PLUGINS was added.

Plugins supporting this features must add themselves to the list of
statically linked plugins based on the setting of WITH_STATIC_PLUGINS.

additionally the macro DECLARE_LAYERMANAGEMENT_PLUGIN was added to
simplify the definition of a plugin entry point for loading.

Signed-off-by: Timo Lotterbach <timo.lotterbach@bmw-carit.de>