LayerManagerService: updated IPlugin, PluginBase
all plugins are now required to provide their API type and version
as well as their name.

This allows loading all plugins using the same mechanism and detect
their type during runtime. The API version can be used to make sure,
that the LayerManagerService supports this verison of the plugin API.

The name of the plugin is used for logging.

Signed-off-by: Timo Lotterbach <timo.lotterbach@bmw-carit.de>