Wayland: fixed build of Wayland Renderers caused by Health Monitoring API change
Health monitoring is updated to detect thread dead-lock by incrementing
thread internal counters. The health monitoring API of Plugins was modified
to request the internal iteration counter value.

This allows the health monitor to check, if the plugin is still working.

This patch enables the build, but does not implement iteration counter
handling in Wayland-based Renderers.

Signed-off-by: Timo Lotterbach <timo.lotterbach@bmw-carit.de>