ScreenDump: check for valid screenId and set screenId for layer
If rendering plug-in supports several screens we cannot assume
only screen 0 as valid screenId.
Also if the layer is assigned to some screen, containing screenId
should be set.
One layer could also be assigned to more than one
screen at the same time, in this case only the last screenId is set

Signed-off-by: Eugen Friedrich <efriedrich@de.adit-jv.com>